Refactor monitor package
