+++ monitor/monitor.go
@@ -1,4 +1,4 @@
-package app
+package monitor
 
 import (
 	"context"
@@ -6,8 +6,6 @@ import (
 
 	"github.com/celer-network/sgn/mainchain"
 	"github.com/celer-network/sgn/utils"
-	"github.com/celer-network/sgn/x/global"
-	"github.com/celer-network/sgn/x/subscribe"
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/ethereum/go-ethereum/core/types"
 )
@@ -51,12 +49,7 @@ func (m *EthMonitor) monitorBlockHead() {
 		case err := <-sub.Err():
 			log.Printf("SubscribeNewHead err", err)
 		case header := <-headerChan:
-			msg := global.NewMsgSyncBlock(header.Number.Uint64(), m.transactor.Key.GetAddress())
-			_, err := m.transactor.BroadcastTx(msg)
-			if err != nil {
-				log.Printf("SyncBlock err", err)
-				return
-			}
+			m.handleNewBlock(header)
 		}
 	}
 }
@@ -78,24 +71,4 @@ func (m *EthMonitor) monitorIntendSettle() {
 			m.handleIntendSettle(intendSettle)
 		}
 	}
-}
-
-func (m *EthMonitor) handleIntendSettle(intendSettle *mainchain.CelerLedgerIntendSettle) {
-	request, err := subscribe.CLIQueryRequest(m.cdc, m.transactor.CliCtx, subscribe.StoreKey, intendSettle.ChannelId[:])
-	if err != nil {
-		log.Printf("Query request err", err)
-		return
-	}
-
-	if intendSettle.SeqNums[request.PeerFromIndex].Uint64() >= request.SeqNum {
-		log.Printf("Ignore the intendSettle event due to larger seqNum")
-		return
-	}
-
-	tx, err := m.ethClient.Ledger.IntendSettle(m.ethClient.Auth, request.SignedSimplexStateBytes)
-	if err != nil {
-		log.Printf("Tx err", err)
-		return
-	}
-	log.Printf("Tx detail", tx)
 }
